serial: msm_geni_serial: Disable Flow before stopping Rx
Add manual flow control to disable and enable flow as per the
driver need. Basically at the time of stop rx, we do not expect
peer device to send any data on RX line. In some cases, peer
device is not in sync with client driver and UART driver reads
unexpected data. This change just prevents the same.

Change-Id: I2bdc70bc68198f441c8e7051c799353abcb4c41c
Signed-off-by: Mukesh Kumar Savaliya <msavaliy@codeaurora.org>